Move NullAsset to SDK
